TAA-440: MTKC timeout
Avoid MTKC hangs by timing out on delete in a reasonable time.
The container would timeout on the agent much later, whereas 10m
is a more than sufficient time for MTKC to remove the testbench.

Change stage name, drop 'on Error' to avoid confusion albeit it
is an error handling stage.